Working on Toilet Task
Create a target which when the player hits a score goes up. Also added sound sources that play when scoring (just need to add the sound effects in the audio source). I need to create a score UI and a completion of the task when your score is at a number.  Additionally, i need to work on my other tasks mainly the electrical box and a snack task.